Changing ptr::write_bytes to intrinsics::volatile_set_memory
write_bytes is insufficient where security matters because the compiler
can decide to optimize the instruction out if it seems unnecessary to
it. Therefore volatile_set_memory is used because the volatile-flag
indicates for LLVM that this instruction must happen and is not allowed
to remove.

Moreover rc_weak is stable now, therefore the feature is removed in lib.rs